+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to [diagram-js](https://github.com/bpmn-io/diagram-js) are d
 
 _**Note:** Yet to be released changes appear here._
 
+* `FEAT`: allow `Canvas#findRoot` to find shared root for multiple elements ([#1075](https://github.com/bpmn-io/diagram-js/pull/1075))
+
 ## 15.19.0
 
 _Improves rendering and editing of larger diagrams significantly._